optimize for timer

+++ timer.sh
@@ -1,43 +1,47 @@
 #!/bin/bash
-export MY_HOME=${HOME}
-export BTASK_LIST="master,mdat,ncat,xfer"
-source ${MY_HOME}/.bashrc
+TIMER_RUNDIR=/var/run/timer
+if [ -f ${TIMER_RUNDIR}/timerc ];then
+    source ${TIMER_RUNDIR}/timerc
 
-if can_access "${MY_HOME}/.timerc";then
-    source ${MY_HOME}/.timerc
-else
-    exit 0
-fi
+    export BTASK_LIST="master,mdat,ncat,xfer"
+    source ${MY_HOME}/.bashrc
 
-if can_access "${TEST_SUIT_ENV}";then
-    source ${TEST_SUIT_ENV} 
+    if can_access "${MY_HOME}/.timerc";then
+        source ${MY_HOME}/.timerc
+    else
+        exit 0
+    fi
 
-    if can_access "${ISCSI_APP_LOG}";then
-        logsize=$(file_size "${ISCSI_APP_LOG}")
+    if can_access "${TEST_SUIT_ENV}";then
+        source ${TEST_SUIT_ENV} 
+
+        if can_access "${ISCSI_APP_LOG}";then
+            logsize=$(file_size "${ISCSI_APP_LOG}")
+            if [ -z "${logsize}" ];then
+                logsize=0
+            fi
+
+            maxsize=$((600*1024*1024))
+            if (( logsize > maxsize ));then
+                date_time=$(date '+%Y%m%d-%H%M%S')
+                cp -f ${ISCSI_APP_LOG} ${ISCSI_APP_LOG}.${date_time}
+                #truncate -s 0 ${ISCSI_APP_LOG}
+                cat /dev/null > ${ISCSI_APP_LOG}
+            fi
+        fi
+    fi
+
+    if can_access "${BASHLOG}";then
+        logsize=$(file_size "${BASHLOG}")
         if [ -z "${logsize}" ];then
             logsize=0
         fi
 
-        maxsize=$((600*1024*1024))
+        maxsize=$((300*1024*1024))
         if (( logsize > maxsize ));then
             date_time=$(date '+%Y%m%d-%H%M%S')
-            cp -f ${ISCSI_APP_LOG} ${ISCSI_APP_LOG}.${date_time}
-            #truncate -s 0 ${ISCSI_APP_LOG}
-            cat /dev/null > ${ISCSI_APP_LOG}
+            cp -f ${BASHLOG} ${BASHLOG}.${date_time}
+            cat /dev/null > ${BASHLOG}
         fi
-    fi
-fi
-
-if can_access "${BASHLOG}";then
-    logsize=$(file_size "${BASHLOG}")
-    if [ -z "${logsize}" ];then
-        logsize=0
-    fi
-
-    maxsize=$((300*1024*1024))
-    if (( logsize > maxsize ));then
-        date_time=$(date '+%Y%m%d-%H%M%S')
-        cp -f ${BASHLOG} ${BASHLOG}.${date_time}
-        cat /dev/null > ${BASHLOG}
     fi
 fi
